+++ tests/test_getters.py
@@ -1,3 +1,4 @@
+
 from django.contrib.auth.models import Group, User
 from django.test import TestCase