+++ useful/django/db.py
@@ -1,3 +1,4 @@
+
 from django.db import connection