+++ useful/django/getters.py
@@ -1,3 +1,4 @@
+
 from collections import defaultdict
 
 from django.db import connection