+++ useful/django/readonly_admin.py
@@ -1,3 +1,4 @@
+
 from django.contrib import admin
 from django.core.exceptions import PermissionDenied