+++ useful/django/urlpatterns.py
@@ -2,10 +2,7 @@
 from functools import wraps
 from functools import WRAPPER_ASSIGNMENTS
 import re
-try:
-    from urllib.parse import urlparse
-except ImportError:     # Python 2
-    from urlparse import urlparse
+from urllib.parse import urlparse
 
 from django.conf import settings
 from django.urls import re_path
@@ -16,19 +13,14 @@ from django.utils.encoding import force_str
 from sys import version_info
 
 
-PY2 = version_info[0] == 2
-
-
 def available_attrs(fn):
     """
     Return the list of functools-wrappable attributes on a callable.
     This is required as a workaround for http://bugs.python.org/issue3445
     under Python 2.
+    TODO: still relevant in py3?
     """
-    if PY2:
-        return tuple(a for a in WRAPPER_ASSIGNMENTS if hasattr(fn, a))
-    else:
-        return WRAPPER_ASSIGNMENTS
+    return WRAPPER_ASSIGNMENTS
 
 
 class UrlPatterns(list):
@@ -199,10 +191,7 @@ class UrlPatterns(list):
                 _wrapped.can_url_perms_compiled = compiled_anded_perms
 
             # Assigning different name to avoid changing the nonlocal variable.
-            if PY2:
-                url_name = view_func.func_name if name is () else name
-            else:
-                url_name = view_func.__name__ if name is () else name
+            url_name = view_func.__name__ if name is () else name
             self.append(re_path(regex, _wrapped, kwargs=kwargs, name=url_name))
 
             return _wrapped